bff/src/services/holdingService.ts
import { getPrismaClient } from '../utils/database';
import { CreateHoldingDTO, UpdateHoldingDTO, HoldingDetails } from '../models/Holding';
import * as stockService from './stockService';
import * as quoteService from './quoteService';
import { HoldingRepository } from '../../../db/repositories/HoldingRepository';
import { TransactionRepository } from '../../../db/repositories/TransactionRepository';

// Initialize repositories
const prisma = getPrismaClient();
let holdingRepository = new HoldingRepository(prisma);
let transactionRepository = new TransactionRepository(prisma);

// For testing: allow repository injection
export const setHoldingRepository = (repo: any) => {
  holdingRepository = repo;
};

export const setTransactionRepository = (repo: any) => {
  transactionRepository = repo;
};

// Helper function to map DB Holding to API response
const mapDBHoldingToDetails = async (dbHolding: any): Promise<HoldingDetails> => {
  const stock = await stockService.getStockByISIN(dbHolding.ISIN);
  const quotes = await quoteService.getLatestQuotes([dbHolding.ISIN]);
  
  const currentPrice = quotes[0]?.price || 0;
  const totalValue = currentPrice * dbHolding.QUANTITY;

  // Calculate gain/loss using transaction history
  const transactions = await transactionRepository.findByHolding(dbHolding.HOLDINGS_ID);
  const totalCost = await transactionRepository.getTotalValue(dbHolding.HOLDINGS_ID);
  const gainLoss = totalValue - Number(totalCost);
  const gainLossPercentage = Number(totalCost) > 0 ? (gainLoss / Number(totalCost)) * 100 : 0;

  return {
    HOLDINGS_ID: dbHolding.HOLDINGS_ID,
    PORTFOLIOS_ID: dbHolding.PORTFOLIOS_ID,
    ISIN: dbHolding.ISIN,
    QUANTITY: dbHolding.QUANTITY,
    START_DATE: dbHolding.START_DATE,
    END_DATE: dbHolding.END_DATE,
    stock: {
      symbol: stock?.symbol || '',
      name: stock?.name || '',
      currency: stock?.currency || 'USD'
    },
    currentPrice,
    totalValue,
    gainLoss,
    gainLossPercentage
  };
};

export const createHolding = async (
  holdingData: CreateHoldingDTO
): Promise<HoldingDetails> => {
  try {
    // First verify the stock exists
    const stock = await stockService.getStockByISIN(holdingData.ISIN);
    if (!stock) {
      throw new Error('Stock not found');
    }

    // Create the holding using repository
    const dbHolding = await holdingRepository.create({
      HOLDINGS_ID: '', // Will be generated
      PORTFOLIOS_ID: holdingData.PORTFOLIOS_ID,
      ISIN: holdingData.ISIN,
      QUANTITY: holdingData.QUANTITY,
      START_DATE: new Date(),
      END_DATE: null
    });

    // Create initial transaction using repository
    await transactionRepository.create({
      TRANSACTIONS_ID: '', // Will be generated
      HOLDINGS_ID: dbHolding.HOLDINGS_ID,
      BUY: true, // Initial transaction is always a buy
      AMOUNT: holdingData.QUANTITY,
      PRICE: holdingData.PRICE,
      TRANSACTION_TIME: new Date(),
      COMMISSION: 0,
      BROKER: 'SYSTEM'
    });

    return await mapDBHoldingToDetails(dbHolding);
  } catch (error) {
    if (error instanceof Error) {
      throw error;
    }
    throw new Error('Failed to create holding');
  }
};

export const getHoldingById = async (
  holdingId: string
): Promise<HoldingDetails | null> => {
  const holding = await holdingRepository.findById(holdingId);

  if (!holding) {
    return null;
  }

  return await mapDBHoldingToDetails(holding);
};

export const getHoldingsByPortfolioId = async (
  portfolioId: string
): Promise<HoldingDetails[]> => {
  const holdings = await holdingRepository.findActiveByPortfolio(portfolioId);
  return Promise.all(holdings.map(mapDBHoldingToDetails));
};

export const updateHolding = async (
  holdingId: string,
  updateData: UpdateHoldingDTO
): Promise<HoldingDetails> => {
  try {
    if (updateData.QUANTITY === undefined) {
      throw new Error('Quantity is required for update');
    }
    const updatedHolding = await holdingRepository.updateQuantity(holdingId, updateData.QUANTITY);
    return await mapDBHoldingToDetails(updatedHolding);
  } catch (error) {
    if (error instanceof Error) {
      throw error;
    }
    throw new Error('Failed to update holding');
  }
};

export const closeHolding = async (holdingId: string): Promise<void> => {
  try {
    await holdingRepository.closeHolding(holdingId, new Date());
  } catch (error) {
    if (error instanceof Error) {
      throw error;
    }
    throw new Error('Failed to close holding');
  }
};
